- Summary: fix bug : fix issues reported by findbugs
- Bugzilla Bug (s) Resolved:

- Description: fix bug : fix issues reported by findbugs

- Tests Description: Manual test.

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Added:


- Files Edited:


- Files Deleted:

+++ UI/org.eclipse.birt.report.designer.ui/src/org/eclipse/birt/report/designer/internal/ui/swt/custom/TabbedPropertyList.java
@@ -585,7 +585,7 @@ public class TabbedPropertyList extends Canvas
 	 */
 	protected void computeTabsThatFitInComposite( )
 	{
-		tabsThatFitInComposite = Math.round( ( getSize( ).y - 22 )
+		tabsThatFitInComposite = Math.round( ( (float) ( getSize( ).y - 22 ) )
 				/ getTabHeight( ) );
 		if ( tabsThatFitInComposite <= 0 )
 		{
